Modified group assignments and collapsing

+++ matcher.py
@@ -346,15 +346,15 @@ def stilts_crossmatch_N(logger,  path_dictionary: dict) -> pd.DataFrame:
             except:
                 pass
 
-def stilts_internal_match(logger,  catalog_path: Path, batch_n : int = -1) -> Tuple[pd.DataFrame, pd.DataFrame]:
+def stilts_internal_match(logger,  catalog_path: Path, batch_n : int = -1, keys : str = "GroupID", do_centroids: bool = False) -> Tuple[pd.DataFrame, pd.DataFrame]:
         """Run STILTS crossmatch between two catalogs."""
         logger.info(f"Finding groups in {catalog_path.name} via STILTS internal match")
         
         with tempfile.NamedTemporaryFile(suffix='.parquet', delete=False) as tmp_file:
             temp_output = Path(tmp_file.name)
-        
-        with tempfile.NamedTemporaryFile(suffix='.parquet', delete=False) as tmp_file_centroid:
-            temp_output_centroid = Path(tmp_file_centroid.name)
+        if do_centroids:
+            with tempfile.NamedTemporaryFile(suffix='.parquet', delete=False) as tmp_file_centroid:
+                temp_output_centroid = Path(tmp_file_centroid.name)
 
         try:
             cmd = [
@@ -380,30 +380,39 @@ def stilts_internal_match(logger,  catalog_path: Path, batch_n : int = -1) -> Tu
                 text=True,
                 check=True
             )
-            
-            cmd_centroid = [
-                'java', '-jar', STILTS,
-                '-stilts', 'tgroup',
-                f"in={temp_output}", 'ifmt=parquet',
-                'keys=GroupID',
-                f"aggcols=0;count {CROSSMATCH['col1_ra']};mean {CROSSMATCH['col1_dec']};mean",
-                'omode=out',
-                f'out={temp_output_centroid}', 'ofmt=parquet'
-            ]
-            
-            result_centroid = subprocess.run(
-                cmd_centroid,
-                capture_output=True,
-                text=True,
-                check=True
-            )
 
+            df_centroid = None
             df = pd.read_parquet(temp_output)
-            df_centroid = pd.read_parquet(temp_output_centroid)
+            # Find where GroupID is NaN (non-matched)
+            nan_mask = df['GroupID'].isna()
+            # Generate a sequence that continues from max group and assign to NaNs
+            max_group = df['GroupID'].max()
+            df.loc[nan_mask, 'GroupID'] = range(max_group + 1, nan_mask.sum() + 2)
+            
+            if do_centroids:
+                cmd_centroid = [
+                    'java', '-jar', STILTS,
+                    '-stilts', 'tgroup',
+                    f"in={temp_output}", 'ifmt=parquet',
+                    f'keys={keys}',
+                    f"aggcols=0;count {CROSSMATCH['col1_ra']};mean {CROSSMATCH['col1_dec']};mean",
+                    'omode=out',
+                    f'out={temp_output_centroid}', 'ofmt=parquet'
+                ]
+                
+                result_centroid = subprocess.run(
+                    cmd_centroid,
+                    capture_output=True,
+                    text=True,
+                    check=True
+                )
+                df_centroid = pd.read_parquet(temp_output_centroid)
+                if batch_n >= 0:
+                    df_centroid["batch"] = batch_n
+            
             #df_centroid.drop([CROSSMATCH['col1_ra'], CROSSMATCH['col1_dec']], axis=1, inplace=True)
             # must change this to use the mean values as new RA and Dec
-            if batch_n >= 0:
-                df_centroid["batch"] = batch_n
+
             logger.info(f"Crossmatch completed: {len(df)} rows, {len(df.columns)} columns")
             
             return df, df_centroid
@@ -414,7 +423,8 @@ def stilts_internal_match(logger,  catalog_path: Path, batch_n : int = -1) -> Tu
         finally:
             try:
                 temp_output.unlink()
-                temp_output_centroid.unlink()
+                if do_centroids:
+                    temp_output_centroid.unlink()
             except:
                 pass
 
@@ -446,7 +456,7 @@ def match_list_of_files(logger, paths, idx):
             batch_df = pd.concat(batch_dfs, ignore_index=True)
             batch_df.to_parquet(current_temp_file, index=False)
             logger.info(f"Joined batch {i//batch_size + 1}")
-            full_output, collapsed_output = stilts_internal_match(logger, Path(current_temp_file), i)
+            full_output, collapsed_output = stilts_internal_match(logger, Path(current_temp_file), i, do_centroids=True)
             subsets.append(full_output)
             subset_centroids.append(collapsed_output)
             Path(current_temp_file).unlink()
@@ -459,7 +469,7 @@ def match_list_of_files(logger, paths, idx):
         # Here stilts internal match, all_rows should contain a column with groups
         # while current_result a single row per group. Perhaps include statistics?
         # all_rows = stilts_internal_match(logger, Path(current_temp_file))
-        _, current_result = stilts_internal_match(logger, Path(final_temp_file))
+        current_result, _ = stilts_internal_match(logger, Path(final_temp_file))
 
 
     except Exception as e:
